Cache Performance Fix, Tag Search
The performance of the hash map in the factory base was abysmal for
large bulk operations, so was refactored.  Includes data search by tag,
though no convenient way to tag the data at the moment.

+++ src/AccountManagerDataAccess/src/main/java/org/cote/accountmanager/data/factory/AddressFactory.java
@@ -210,7 +210,7 @@ public class AddressFactory extends NameIdGroupFactory {
 
 	}
 */
-	public List<AddressType>  getAddressList(QueryField[] fields, int startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
+	public List<AddressType>  getAddressList(QueryField[] fields, long startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
 	{
 		return getPaginatedList(fields, startRecord, recordCount, organization);
 	}
@@ -220,7 +220,7 @@ public class AddressFactory extends NameIdGroupFactory {
 	}
 	
 	
-	public List<AddressType> searchAddresss(String searchValue, int startRecord, int recordCount, DirectoryGroupType dir) throws FactoryException{
+	public List<AddressType> searchAddresss(String searchValue, long startRecord, int recordCount, DirectoryGroupType dir) throws FactoryException{
 	
 		ProcessingInstructionType instruction = null;
 		if(startRecord >= 0 && recordCount >= 0){
