+++ src/AccountManagerDataAccess/src/main/java/org/cote/accountmanager/data/factory/OperationFactory.java
@@ -213,7 +213,7 @@ public class OperationFactory extends NameIdGroupFactory {
 		return convertList(lst);
 
 	}
-	public List<OperationType>  getOperationList(QueryField[] fields, int startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
+	public List<OperationType>  getOperationList(QueryField[] fields, long startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
 	{
 		return getPaginatedList(fields, startRecord, recordCount, organization);
 	}
