+++ src/AccountManagerDataAccess/src/main/java/org/cote/accountmanager/data/factory/RuleFactory.java
@@ -230,7 +230,7 @@ public class RuleFactory extends NameIdGroupFactory {
 
 	}
 	
-	public List<RuleType>  getRuleList(QueryField[] fields, int startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
+	public List<RuleType>  getRuleList(QueryField[] fields, long startRecord, int recordCount, OrganizationType organization)  throws FactoryException,ArgumentException
 	{
 		return getPaginatedList(fields, startRecord, recordCount, organization);
 	}
