+++ src/AccountManagerDataAccess/src/main/java/org/cote/accountmanager/data/factory/SessionFactory.java
@@ -530,11 +530,11 @@ public class SessionFactory extends FactoryBase {
 		return out_list;
 	}
 	
-	public ProcessingInstructionType getPagingInstruction(int startIndex)
+	public ProcessingInstructionType getPagingInstruction(long startIndex)
 	{
 		return getPagingInstruction(startIndex, defaultPageSize);
 	}
-	public ProcessingInstructionType getPagingInstruction(int startIndex, int recordCount)
+	public ProcessingInstructionType getPagingInstruction(long startIndex, int recordCount)
 	{
 		ProcessingInstructionType instruction = new ProcessingInstructionType();
